debian/rules: adding hack to preserve correct config.status
This commit is just a hack to get a working file $(buildfolder)/config.status.
The Python magic from Mozilla isn't preparing a correct value for the
variable VISIBILITY_FLAGS.

For some reason this variable is expaned to

   -Isid/system_wrappers

while it should expanded to

  -I/$(CURDIR)/obj-thunderbird/config/system_wrappers

This commit is just a hack. :-(
We'd need to go down into the Mozilla magic and figure out how
config.status is created now and what's going wrong.