debian/control: add missing packages for *-ast language
It seems we've forgotten to proper add the l10n packages for
iceowl-extension in the past while it was introduced by Mozilla, the
l10n package iceowl-l10n-ast and the new package thunderbird-l10n-ast
are exist.
We have now to add the missed l10n packages for lightning and for the
transitional package iceowl-extension.