debian/rules: install icudt5*.dat file more flexible
If we need to use the pre-shipped icu18n stuff a file icudt5*.dat will be
installed by upstream into /usr/lib/thunderbird. Commit
5b04b3275513ab139b53ca742c6ac4779867c654 already try to install that
file but a build with the system library libicu-dev (so without a
pre-shipped icu18n) will cause to break the build.

This commit will enable a mostly flexible solution which will cover both
cases without a need to change something inside debian/rules or the dh
sequencer files.
The package thunderbird-dev is also needing a extra symlink so xpcshell
can work if a pre-shipped icudt5*.dat file is around.