Icedove2Thunderbird: add opt out for dialogue pop-up
Adding a possibility for system administrators to switch of the dialogue
window about the profile migration Icedove to Thunderbird.

The wrapper script is now checking the existence of a file
'/etc/thunderbird/no_migration_popup'. If this file is available no
information pop-up about the automatic profile migration will be
displayed if a user hasn't done the migration already.