debian/rules: build icudt5*.dat on our own if needed
In case the libicu-dev package can't satisfy the build depends of
Thunderbird we need to use the internal icu sources controlled by the
configure option '--without-system-icu'.
Unfortunately Mozilla isn't building then the needed file
icudt5*{l,b}.dat if the configuration for Thunderbird is tuned for using
the internal icu sources. Additionally Mozilla is currently only shipping
a file icudt5*l.dat which is usable on little endian platforms only. This
isn't a problem as long we are on platforms with little endianness like
amd64, arm64, armel, armhf, alpha, hurd, i386, ia64, and kfreebsd-* as
the precompiled binary file is build for little endian.
On Debian we also have hppa, m68k, mips, powerpc{spe}, ppc64, sparc{64}
and s390x which need a file icudt5*b.dat because those are platforms with
big endianness.

This commit checks the usage of the option '--without-system-icu' and is
building the icudt5*.dat file afterwards if needed.

This commit isn't needed if Thunderbird is build without intl-api
support ('--without-intl-api') which we probably don't want to do.