mozconfig.default: use proper disabled options
While options that are commented out are treated as "use internal
version" we should use the explicit enabled/disabled options for such
additional components.
This helps to make it easier to determine the configure status of such
components, we can simpler check for "--with-system-foo" e.g. and can
easy decide what to do if needed.