use icudt??{b,l}.dat depending on architecture
Upstream doesn't really care about big endian platforms and so also don't
care about a needed file icudt??b.dat on big endian platforms. Changing
the m4 macro for ICU so the needed file will be depended on the build
platform exported and let later the configure run automatically use the
correct name for the ICU *.dat file so we not only depend on the
prebuild *.dat file for little endian.

The `case foo in ...` loop could be obviously more compressed with only
one case entry for all architectures which are all big endian.

---%<---
    case $host in
       dnl big endian architectures
       hppa-*|m68k-*|...)
--->%---

For more readability the more exclusive extra list of all possible
architectures with big endianness in Debian was choosen.