debian/control: change maintainer
Christoph stated he hasn't enough time since a few months to maintain
the Thunderbird packages well and wont have time to do so in the near
future.
As I've done mostly the maintaining and packaging part for quite a while
we decided it's probably best to just switch the maintainership for
Thunderbird packages now as this reflects the real state.
If other people want helping in maintaining Thunderbird please don't
hesitate to contact me or Guido!

Thanks Christoph for doing all the packaging stuff since 3.0~rc1-1 (from Thu,
03 Dec 2009) in the last 7 years! You are welcome if you want to step in
again.