Disable the AppArmor profile by default
since AppArmor was enabled by default in sid 9 days ago, it's become
obvious that the AppArmor policy we ship for Thunderbird simply breaks
too many use cases, and there seems to be no way to fix that while
providing meaningful confinement: for example, see #882048 and
by default and let users enable it themselves if they are fine with
dealing with the fallout. That's a bit sad, but it's a much better
output than seeing people burning themselves and getting used to
disabling AppArmor entirely on their system, i.e. losing the benefits
of the other, working profiles we ship.

Closes: #882672, #882218, #882048