calendar-google-provider: tweak installation a bit
Since the change of the plugin inclusion model the way mozilla-devscript
is providing plugins as xul-ext-* seems not to work for
calendar-google-provider anymore. Linking the application id from the
thunderbird extension folder to the content of the package makes the
package functional under Thunderbird again.
This commit is a hackish solution, but some where we need to start. For
now I have no deeper knowledge about the concept of webextension and how
to map this into the packaging of GDataProvider.