d/mozconfig.default: don't use nspr and nss from system
The commit 0aa2546cc083286243bae55833ed22690cb167e7 has missed the
removing of the disabled options for nspr and nss

---%<---
ac_add_options --without-system-nspr   # Debian version to old in unstable
ac_add_options --without-system-nss    # Debian version to old in unstable
--->%---

But due the later explicit enabling of this configure option this wasn't
a problem as the commit wanted to enable the usage of both system
package.
But now TB 60.0b1 needs versions that are not packaged yet and we need
to use the versions again from the upstream source and we remove both
compiler options again.