apparmor: allow access to @{HOME}/.gnupg/tofu.db
From time to time a gpg process consuming 100% of one cpu. A Debian
fellow mentioned me that maybe AppArmor could be the cause, so I dug
into it a little bit and found that AppArmor is denying the access to
gpg to the file ~/.gnupg/tofu.db.
The solution is to allow the access to that file.

Closes: #894907