add entries to README and NEWS for thunderbird
The handling off the l10n stuff has changed for Thunderbird starting with
version 59. Due this we can't provide a automatic UI l10n usage after
the installation of the l10n packages.
Adding some notes for users so they can re-enable at least the l10n
language for the UI of thunderbird, for lightning we haven't found a
solution or workaround yet.