create-lightning-l10n: adjust folder structure
The XPI archives for the localized Ligthning extension containing now
all needed folders and files within the folder
chrome/{calendar,lightning}-$locale which is all we need for the l10n
packages.
Modifiying the created structure for the additional tarball so we can
later simply copy the content into the binary package.