mozconfig.default: disable webrtc build and inclusion
Thunderbird has no use case for WebRTC, real-time, peer-to-peer, media
exchange; via a discovery and negotiation process called signaling for
real-time exchange of audio, video, and data.
The only place this could be useful would have support of WebRTC for the
chat client, but so far it's still not implemented. So we will switch of
the build of webrtc for now, if later needed we can simply turn it on
again.

See
https://bugzilla.mozilla.org/show_bug.cgi?id=1018060