Revert "apparmor: allow access to @{HOME}/.gnupg/tofu.db"
This reverts commit 840cbc8a510e5981e3759e89c1edb8cff5e659a1.

The same change was done already in commit
4656ebf88d5dc92964b14da54198216e4182a110.

Besides, the process to update this AppArmor profile is to first get the changes
upstream (which includes a review by other Debian/AppArmor people) and then to
copy the updated upstream profile to our Vcs-Git. Changes applied directly in
our Vcs-Git will be overwritten next time that copy is done.