adjust the configuration of the built
As the subfolder mozilla/ isn't existing anymore and the mail
application Thunderbird itself has been move into comm/mail/ the
configuration for the built needs to be changed too.

The configuration (as also later the built) is controled by the 'mach'
script which is a Python script that is wraping all the needed
configuration steps.