c-l-l10n-t.sh: adjust the path to the python helper
After the changes to the folder structure the helper shell script need
to be modified to use the makeversion.py Python script again.
Unfortunately upstream hasn't fixed the build of the lightning extension
after moving over to buildcluster so we need to cheat a bit for the
Lightning l10n files for now and use the previous files from an earlier
version.