d/c-l-l-t.sh: update to use thunderbird archives
Upstream isn't providing dedicated Lightning packagaes anymore and thus
we can't use the previous way to collect the l10n data for the Lightning
AddOn.
Insteat we need to get now all the prebuild Thunderbird packages for the
various languages and extract the l10n string from these archives.

The changes to the script are a bit hacky and don't work if you dont use
the option '-e' at least. The script needs to be migrated into a Python
script to get all the files in one go.