d/rules: use MOZ_LANGPACK_ID instead of hard coding
We don't build the l10n packages for thunderbird on our own and use
instead data from the pre-compiled binaries from Mozilla. This is
because the TB build system isn't building the language files, they are
build outside within the taskcluster environment while Mozilla is
creating the binary packages.

The makefile comm/mail/locales/Makefile.in is defining the used org id
in the binary package later by using the variable MOZ_LANGPACK_EID. If
the domian name in the this variable doesn't match the the
application.id in the manifest.json file the created l10n package won't
be usable later.

This commit is using the top level domain name in from the variable
MOZ_LANGPACK_EID so we can build xpi files based on the use domain name.