d/rules: ensure l10n MOZ_LANGPACK_ID matches variable from makefile
To get usable thunderbirs-l10n packages later we need to check if the
variable MOZ_LANGPACK_ID, which is pulled from
comm/mail/locales/Makefile.in, is equal to the used domain name taken from
manifest.json file of each language. If both are not the same break the
build as this needs to be fixed before we can build the language
packages.

Closes: #906176