locale: use 'intl.locale.requested' correctly
The new environment intl.locale.requested is more than a hard
environment that is setting up a desired language for the UI of
Thunderbird.
Usee with an empty string it does the same as the old environment
intl.locale.matchOS, picking up the system locale. So adding this new
setting "correctly" to /etc/thunderbird/pref/thunderbird.js remove the
requirement that user need to modify their prefs to use TB with the
desired UI language.

Closes: #908034